+++ local-development.php
@@ -13,8 +13,8 @@
  * Plugin URI:        https://wordpress.org/plugins/local-development
  * Author:            Andy Fragen
  * Author URI:        http://thefragens.com/
- * Description:       Places development notice for plugins or themes that are in local development. Prevents updating of selected plugins and themes. Allows for using a local file server and bypassing the WSOD Shutdown Handler.
- * Version:           2.3.0
+ * Description:       Places development notice for plugins or themes that are in local development. Prevents updating of selected plugins and themes. Allows for using a local file server and bypassing the WordPress 5.2 WSOD Shutdown Handler.
+ * Version:           2.4.0
  * Domain Path:       /languages
  * Text Domain:       local-development
  * License:           GNU General Public License v2